% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Principal Component Analysis}
\usage{
PCA(
  df,
  rows = NULL,
  columns = NULL,
  patterns = NULL,
  exclude = T,
  retain = NULL
)
}
\arguments{
\item{df}{A data frame.}

\item{rows}{An optional logical vector matching in
length to the number of rows in 'df' indicating
which rows to use for the PCA.}

\item{columns}{An optional vector of column names
in 'df' to use for the PCA.}

\item{patterns}{An optional vector of character
strings, incomplete patterns to use to identify
either columns to include in or columns to exclude
from the PCA.}

\item{exclude}{Logical; if TRUE, the 'patterns' vector
are strings to identify columns to exclude; otherwise
the 'patterns' vector are strings to identify columns
to include.}

\item{retain}{Description.}
}
\value{
An object of class 'PCA'.
}
\description{
A convenience function that conducts a
prinicipal components analysis (PCA) on
the specified rows and columns of a
data frame using R's base
function \code{\link[stats]{princomp}}.
}
\details{
Once a PCA has been conducted,
prinicipal component scores are
computed for all observations.
}
\examples{
# Example of function
results = PCA( mtcars, patterns = c( 'p', 'r' ) )

}
