% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAandEFA-package.R
\docType{package}
\name{PCAandEFA-package}
\alias{PCAandEFA-package}
\title{Convenience Functions to Run PCA or EFA}
\description{
The \pkg{PCAandEFA} package provides a set of
functions to streamline running a principal
components analysis (PCA) or an exploratory
factor analysis (EFA). A complete  list of the
functions available in the package can be obtained
via the command \code{ls(pos = "package:PCAandEFA")}.
}
